# apps/api/docs/swagger.yaml
basePath: /api/v1
definitions:
...
  github_com_orkai-dev_orkai_apps_api_internal_apidocs.ProblemDetail:
    properties:
      detail:
        type: string
      errors:
        items:
          $ref: '#/definitions/github_com_orkai-dev_orkai_apps_api_internal_apierr.FieldError'
        type: array
      instance:
        type: string
      status:
        type: integer
      title:
        type: string
      type:
        type: string
    type: object
...
  github_com_orkai-dev_orkai_apps_api_internal_apidocs.Template:
    properties:
      category:
        type: string
      compose_yaml:
        type: string
      created_at:
        type: string
      description:
        type: string
      env_schema:
        additionalProperties:
          $ref: '#/definitions/github_com_orkai-dev_orkai_apps_api_internal_model.EnvVar'
        type: object
      id:
        type: string
      logo_url:
        type: string
      min_cpu:
        type: string
      min_memory:
        type: string
      name:
        type: string
      updated_at:
        type: string
    type: object
  github_com_orkai-dev_orkai_apps_api_internal_apidocs.TemplateListResponse:
    properties:
      items:
        items:
          $ref: '#/definitions/github_com_orkai-dev_orkai_apps_api_internal_model.Template'
        type: array
      pagination:
        $ref: '#/definitions/github_com_orkai-dev_orkai_apps_api_internal_httputil.Pagination'
    type: object
...
  github_com_orkai-dev_orkai_apps_api_internal_apierr.FieldError:
    properties:
      field:
        type: string
      message:
        type: string
    type: object
  github_com_orkai-dev_orkai_apps_api_internal_httputil.Pagination:
    properties:
      page:
        type: integer
      per_page:
        type: integer
      total:
        type: integer
      total_pages:
        type: integer
    type: object
...
  github_com_orkai-dev_orkai_apps_api_internal_model.EnvVar:
    properties:
      default:
        type: string
      description:
        type: string
      required:
        type: boolean
      secret:
        type: boolean
    type: object
...
  github_com_orkai-dev_orkai_apps_api_internal_model.Template:
    properties:
      category:
        type: string
      compose_yaml:
        type: string
      created_at:
        type: string
      description:
        type: string
      env_schema:
        additionalProperties:
          $ref: '#/definitions/github_com_orkai-dev_orkai_apps_api_internal_model.EnvVar'
        type: object
      id:
        type: string
      logo_url:
        type: string
      min_cpu:
        type: string
      min_memory:
        type: string
      name:
        type: string
      updated_at:
        type: string
    type: object
...
info:
  contact: {}
  description: REST control plane for orka'i self-hosted PaaS
  title: Orkai API
  version: "1.0"
paths:
...
  /templates:
    get:
      parameters:
      - description: Page number
        in: query
        name: page
        type: integer
      - description: Items per page (max 100)
        in: query
        name: per_page
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/github_com_orkai-dev_orkai_apps_api_internal_apidocs.TemplateListResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/github_com_orkai-dev_orkai_apps_api_internal_apidocs.ProblemDetail'
      security:
      - BearerAuth: []
      summary: List templates
      tags:
      - templates
  /templates/{id}:
    get:
      parameters:
      - description: Template ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/github_com_orkai-dev_orkai_apps_api_internal_apidocs.Template'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/github_com_orkai-dev_orkai_apps_api_internal_apidocs.ProblemDetail'
      security:
      - BearerAuth: []
      summary: Get template
      tags:
      - templates
...
securityDefinitions:
  BearerAuth:
    description: JWT access token or API key (ork_…)
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"